cmd/native-btc/relayer_test.go
package main

import (
	"testing"
	"time"

	"github.com/btcsuite/btcd/chaincfg/chainhash"
	"github.com/btcsuite/btcd/wire"
	"github.com/gonative-cc/relayer/native/database"
)

// Mock Bitcoin RPC client for testing
type MockBitcoinClient struct{}

func (m *MockBitcoinClient) SendRawTransaction(tx *wire.MsgTx, allowHighFees bool) (*chainhash.Hash, error) {
	// Mock implementation
	hash, err := chainhash.NewHashFromStr("test-tx-hash")
	if err != nil {
		return nil, err
	}
	return hash, nil
}

func (m *MockBitcoinClient) Shutdown() {
	// Mock implementation (if needed)
}

func TestRelayerStart(t *testing.T) {
	// Initialize an in-memory database for testing
	err := database.InitDB(":memory:")
	if err != nil {
		t.Fatal(err)
	}

	// Insert some pending transactions
	transactions := []database.Transaction{
		{BtcTxID: 1, RawTx: "tx1-hex", Status: database.StatusPending},
		{BtcTxID: 2, RawTx: "tx2-hex", Status: database.StatusPending},
	}
	for _, tx := range transactions {
		err = database.InsertTransaction(tx)
		if err != nil {
			t.Fatal(err)
		}
	}

	// Create a Relayer with a mock Bitcoin client
	config := Config{
		DatabasePath: ":memory:",
		BitcoinNode:  "mock-node", // Not used with the mock client
	}
	relayer, err := NewRelayer(config)
	if err != nil {
		t.Fatal(err)
	}
	relayer.btcClient = &MockBitcoinClient{} // Use the mock client

	// Start the relayer in a separate goroutine
	go relayer.Start()

	// Wait for a short time to allow the relayer to process transactions
	time.Sleep(time.Second * 2)

	// Stop the relayer
	relayer.Stop()

	// Add assertions to check if transactions were broadcasted and updated
	// For example, you can retrieve transactions from the database and
	// check if their status has changed to "broadcast".
}
